Replaced FloatField with DecimalField for financial amounts.

+++ plugins/revenue/models.py
@@ -30,8 +30,8 @@ from shotserver04.common import granular_update
 class UserPayment(models.Model):
     user = models.ForeignKey(User, raw_id_admin=True)
     currency = models.CharField(max_length=3)
-    amount = models.FloatField()
-    euros = models.FloatField()
+    amount = models.DecimalField(max_digits=7, decimal_places=2)
+    euros = models.DecimalField(max_digits=7, decimal_places=2)
     paid = models.DateTimeField()
 
     class Admin:
@@ -61,8 +61,8 @@ class UserDonation(models.Model):
     user = models.ForeignKey(User, raw_id_admin=True)
     non_profit = models.ForeignKey(NonProfit)
     currency = models.CharField(max_length=3)
-    amount = models.FloatField()
-    euros = models.FloatField()
+    amount = models.DecimalField(max_digits=7, decimal_places=2)
+    euros = models.DecimalField(max_digits=7, decimal_places=2)
     donated = models.DateTimeField()
 
     class Admin:
